Refactor UI elements to their own directory
