+++ Phoenix/Client/Include/Client/UI/CMakeLists.txt
@@ -1,0 +1,14 @@
+set(currentDir ${CMAKE_CURRENT_LIST_DIR})
+set(Headers
+	${Headers}
+
+	${currentDir}/Crosshair.hpp
+	${currentDir}/DebugOverlay.hpp
+	${currentDir}/EscapeMenu.hpp
+	${currentDir}/GameTools.hpp
+	${currentDir}/HUD.hpp
+	${currentDir}/InventoryUI.hpp
+	${currentDir}/SplashScreen.hpp
+
+	PARENT_SCOPE
+	)
