+++ Phoenix/Client/Source/CMakeLists.txt
@@ -1,26 +1,18 @@
 add_subdirectory(Graphics)
 add_subdirectory(Game)
 add_subdirectory(Voxels)
+add_subdirectory(UI)
 
 set(currentDir ${CMAKE_CURRENT_LIST_DIR})
 set(Sources
 	${Sources}
 
-	${currentDir}/InputQueue.cpp
-
 	${currentDir}/AudioRegistry.cpp
 	${currentDir}/Client.cpp
-	${currentDir}/SplashScreen.cpp
-	${currentDir}/Crosshair.cpp
 	${currentDir}/Game.cpp
-	${currentDir}/EscapeMenu.cpp
-	${currentDir}/HUD.cpp
 	${currentDir}/InputMap.cpp
+	${currentDir}/InputQueue.cpp
 	${currentDir}/Network.cpp
-
-	${currentDir}/DebugOverlay.cpp
-	${currentDir}/GameTools.cpp
-	${currentDir}/InventoryUI.cpp
 
 	${currentDir}/Main.cpp
