+++ Phoenix/Client/Source/UI/CMakeLists.txt
@@ -1,0 +1,14 @@
+set(currentDir ${CMAKE_CURRENT_LIST_DIR})
+set(Sources
+	${Sources}
+
+	${currentDir}/Crosshair.cpp
+	${currentDir}/DebugOverlay.cpp
+	${currentDir}/EscapeMenu.cpp
+	${currentDir}/GameTools.cpp
+	${currentDir}/HUD.cpp
+	${currentDir}/InventoryUI.cpp
+	${currentDir}/SplashScreen.cpp
+
+	PARENT_SCOPE
+	)
