+++ Phoenix/Client/Source/UI/GameTools.cpp
@@ -26,69 +26,40 @@
 // ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 
-#include <Client/DebugOverlay.hpp>
+#include <Client/UI/GameTools.hpp>
 
-#include <Client/Graphics/ImGuiExtensions.hpp>
+#include <Common/Actor.hpp>
+#include <Common/Position.hpp>
+
 #include <imgui.h>
-
-#include <glad/glad.h>
 
 using namespace phx::client;
 using namespace phx;
 
-DebugOverlay::DebugOverlay() : Overlay("DebugOverlay")
+GameTools::GameTools(bool* followCam, entt::registry* registry,
+                     entt::entity player)
+    : Overlay("GameTools"), m_followCam(followCam), m_registry(registry),
+      m_player(player)
 {
-	SDL_DisplayMode current;
-
-	for (int i = 0; i < SDL_GetNumVideoDisplays(); ++i)
-	{
-		if (SDL_GetCurrentDisplayMode(i, &current) == 0)
-		{
-			if (current.refresh_rate > m_maxSampleRate)
-				m_maxSampleRate = current.refresh_rate;
-		}
-	}
 }
 
-void DebugOverlay::onAttach() {}
+void GameTools::onAttach() {}
 
-void DebugOverlay::onDetach() {}
+void GameTools::onDetach() {}
 
-void DebugOverlay::onEvent(events::Event& e) {}
+void GameTools::onEvent(events::Event& e) {}
 
-void DebugOverlay::tick(float dt)
+void GameTools::tick(float dt)
 {
-	ImGui::SetNextWindowPos({50, 50}, ImGuiCond_Once);
 	ImGui::Begin("Phoenix");
-
-	if (ImGui::CollapsingHeader("Graphics Information"))
+	if (ImGui::CollapsingHeader("Game Tools"))
 	{
-		if (ImGui::Checkbox("Wireframe", &m_wireframe))
-		{
-			if (m_wireframe)
-				glPolygonMode(GL_FRONT_AND_BACK, GL_LINE);
-			else
-				glPolygonMode(GL_FRONT_AND_BACK, GL_FILL);
-		}
+		ImGui::Checkbox("Follow Camera", m_followCam);
 
-		ImGui::Text("Frame Time: %.2f ms/frame\n", dt * 1000.f);
-		ImGui::Text("FPS: %d\n", static_cast<int>(1.f / dt));
-
-		ImGui::SliderInt("Debug Sample Rate", &m_sampleRate, 1, 60);
-		ImGui::Checkbox("Pause Debug Graph", &m_pauseSampling);
-
-		if (m_time % m_sampleRate == 0 && !m_pauseSampling)
-		{
-			ImGui::PlotVariable("Frame Time: ", dt * 1000.f);
-		}
-		else
-		{
-			ImGui::PlotVariable("Frame Time: ", FLT_MAX);
-		}
+		ImGui::Text("X: %f\nY: %f\nZ: %f",
+		            m_registry->get<Position>(m_player).position.x,
+		            m_registry->get<Position>(m_player).position.y,
+		            m_registry->get<Position>(m_player).position.z);
 	}
 	ImGui::End();
-
-	++m_time;
-	if (m_time >= 3600)
-		m_time = 0;
 }
